delete route done

+++ backend/controllers/post.js
@@ -27,12 +27,18 @@ exports.modifyPost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
     Post.findOne({_id: req.params.id})
         .then((post) => {
+            if(post.imageUrl) {
             const filename = post.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 Post.deleteOne({_id: req.params.id})
                     .then(() => res.status(200).json({ message : "Publication supprimée !"}))
                     .catch(error => res.status(400).json({ error }));
-            });
+            })}
+            else{
+                Post.deleteOne({_id: req.params.id})
+                    .then(() => res.status(200).json({ message : "Publication supprimée !"}))
+                    .catch(error => res.status(400).json({ error }));
+            }
         })
         .catch(error => res.status(500).json({ error }));
 }
